move password strength bar from ui components to admin panel
