+++ packages/admin-panel/package.json
@@ -27,6 +27,7 @@
     "react-dom": "^16.13.1",
     "react-hook-form": "^6.8.3",
     "react-icons": "^2.2.7",
+    "react-password-strength-bar": "^0.3.2",
     "react-redux": "^5.0.6",
     "react-router-dom": "^5.2.0",
     "react-select": "^1.2.1",
