+++ packages/ui-components/package.json
@@ -35,7 +35,6 @@
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-hook-form": "^6.0.6",
-    "react-password-strength-bar": "^0.3.2",
     "react-router": "^5.1.2",
     "react-router-dom": "^5.1.2",
     "styled-components": "^5.0.1"
